ファイルロックとテストファイルの追加

+++ ip-mag/src/main/java/org/seasar/ip/mag/u_data/FileManager.java
@@ -8,6 +8,8 @@ public class FileManager {
 	private File main_file; // 操作するファイル
 	private ArrayList<String> file_str_array; // 操作するファイルの中身
 	private int file_index; // 行番号
+	private File lock_file; // ロックファイル
+	private String lock_file_url; // ロックファイルのurl
 	
 	public FileManager() {		
 	}
@@ -16,6 +18,12 @@ public class FileManager {
 	public FileManager(String url) {
 		this.file_url = url;
 		this.main_file = new File(url);
+		this.lock_file = null;
+		this.lock_file_url = 
+				this.file_url.substring(
+						0,
+						this.file_url.lastIndexOf(File.separator) + 1)
+						+ "lock";
 		
 		this.file_index = 0;
 		this.file_str_array = new ArrayList<String>();
@@ -57,16 +65,44 @@ public class FileManager {
 		}
 	}
 	
+	/* ファイルロック */
+	public boolean createLockFile() {
+		boolean result = false;
+		
+		if(this.lock_file == null) {
+			this.lock_file = new File(this.lock_file_url);
+		
+			try {
+				result = this.lock_file.createNewFile();
+			} catch(IOException e) {
+				e.printStackTrace();
+			}
+		}
+		
+		return result;
+	}
+	
+	public void deleteLockFile() {
+		if(this.lock_file != null) {
+			this.lock_file.delete();
+			this.lock_file = null;
+		}
+	}
+	/* ファイルロック */
+	
 	/* ............................... */
 	public void clearFile() {
-		this.file_str_array.clear();
-		this.main_file.delete();
-		this.main_file = new File(this.file_url);
+		if(this.createLockFile()) { 			
+			this.file_str_array.clear();
+			this.main_file.delete();
+			this.main_file = new File(this.file_url);
 		
-		try {
-			this.main_file.createNewFile();
-		} catch(IOException e) {		
-			e.printStackTrace();
+			try {
+				this.main_file.createNewFile();
+				this.deleteLockFile();
+			} catch(IOException e) {		
+				e.printStackTrace();
+			}
 		}
 	}
 	
@@ -77,18 +113,20 @@ public class FileManager {
 	
 	// 格納したファイルデータを書き込む
 	public void flush() {
-		try {
-			PrintWriter pw = new PrintWriter(this.main_file);
-			ArrayList<String> als = this.file_str_array;
-		
-			for(int i = 0; i < als.size(); i++) {
-				pw.println(als.get(i));
+		if(this.createLockFile()) {			
+			try {				
+				PrintWriter pw = new PrintWriter(this.main_file);
+				ArrayList<String> als = this.file_str_array;
+			
+				for(int i = 0; i < als.size(); i++) {
+					pw.println(als.get(i));
+				}
+				pw.close();
+				
+				this.deleteLockFile();
+			} catch(FileNotFoundException e) {
+				e.printStackTrace();
 			}
-			pw.close();
-		} catch(FileNotFoundException e) {
-			e.printStackTrace();
-		} catch(IOException e) {
-			e.printStackTrace();
 		}
 	}
 	/* ............................... */
@@ -109,42 +147,51 @@ public class FileManager {
 	
 	// ファイル末尾に追加
 	public void append(String str) {
-		try {
-			PrintWriter pw = new PrintWriter(
-					new FileWriter(this.main_file, true));
+		if(this.createLockFile()) {
+			try {				
+				PrintWriter pw = new PrintWriter(
+						new FileWriter(this.main_file, true));
+				
+				pw.println(str);
+				pw.close();
+				
+				this.deleteLockFile();
+			} catch(FileNotFoundException e) {	
+				e.printStackTrace();
+			} catch(IOException e) {
+				e.printStackTrace();
+			}
 			
-			pw.println(str);
-			pw.close();
-		} catch(FileNotFoundException e) {	
-			e.printStackTrace();
-		} catch(IOException e) {
-			e.printStackTrace();
+			this.file_str_array.add(str);
 		}
-		
-		this.file_str_array.add(str);
 	}
 	
 	// ファイルの指定行を置換
 	public void replaceLine(int line, String str) {
 		this.file_str_array.set(line - 1, str);
 		
-		try {
-			//元のファイルは削除して新しいファイルを作る
-			this.main_file.delete();
-			this.main_file = new File(this.file_url);	
-			this.main_file.createNewFile();
-			
-			PrintWriter pw = new PrintWriter(this.main_file);
-			ArrayList<String> als = this.file_str_array;
-			
-			for(int i = 0; i < als.size(); i++) {
-				pw.println(als.get(i));
+		
+		if(this.createLockFile()) {
+			try {				
+				//元のファイルは削除して新しいファイルを作る
+				this.main_file.delete();
+				this.main_file = new File(this.file_url);	
+				this.main_file.createNewFile();
+				
+				PrintWriter pw = new PrintWriter(this.main_file);
+				ArrayList<String> als = this.file_str_array;
+				
+				for(int i = 0; i < als.size(); i++) {
+					pw.println(als.get(i));
+				}
+				pw.close();
+				
+				this.deleteLockFile();
+			} catch(FileNotFoundException e) {		
+				e.printStackTrace();
+			} catch(IOException e) {
+				e.printStackTrace();
 			}
-			pw.close();
-		} catch(FileNotFoundException e) {		
-			e.printStackTrace();
-		} catch(IOException e) {
-			e.printStackTrace();
 		}
 	}
 }
